refactor(#67): 어뎁터 패키지 분리
