+++ app/src/main/java/com/mimo/android/core/di/NetworkModule.kt
@@ -20,8 +20,8 @@ object NetworkModule {
     @Provides
     fun provideRetrofit(okHttpClient: OkHttpClient): Retrofit {
         // val apiKey = "http://192.168.0.115:8080/"
-        //val apiKey = "http://192.168.123.112:8080/"
-        val apiKey = "http://192.168.1.9:8080/"
+        // val apiKey = "http://192.168.123.112:8080/"
+        val apiKey = "http://192.168.1.13:8080/"
         return Retrofit.Builder()
             .addConverterFactory(GsonConverterFactory.create(GsonBuilder().setLenient().create()))
             .baseUrl(apiKey)
