+++ app/src/main/java/com/mimo/android/presentation/component/adapter/DiffUtilCallback.kt
@@ -1,4 +1,4 @@
-package com.mimo.android.presentation.util
+package com.mimo.android.presentation.component.adapter
 
 import android.annotation.SuppressLint
 import androidx.recyclerview.widget.DiffUtil
