+++ app/src/main/java/com/mimo/android/presentation/component/adapter/MyPostAdapter.kt
@@ -1,4 +1,4 @@
-package com.mimo.android.presentation.mypage.adapter
+package com.mimo.android.presentation.component.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
@@ -6,11 +6,9 @@ import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import com.mimo.android.databinding.ItemMyPostBinding
 import com.mimo.android.domain.model.PostData
-import com.mimo.android.presentation.util.DiffUtilCallback
-import com.mimo.android.presentation.video.upload.TagListAdapter
 
 class MyPostAdapter : ListAdapter<PostData, MyPostAdapter.MyPostViewHolder>(
-    DiffUtilCallback<PostData>()
+    DiffUtilCallback<PostData>(),
 ) {
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): MyPostViewHolder {
@@ -22,16 +20,14 @@ class MyPostAdapter : ListAdapter<PostData, MyPostAdapter.MyPostViewHolder>(
         holder.bind(getItem(position))
     }
 
-
     class MyPostViewHolder(
-        private val binding: ItemMyPostBinding
+        private val binding: ItemMyPostBinding,
     ) : RecyclerView.ViewHolder(binding.root) {
         fun bind(postData: PostData) {
             binding.apply {
                 val tagListAdapter = TagListAdapter()
                 rcTagList.adapter = tagListAdapter
                 tagListAdapter.submitList(postData.tagList)
-
             }
         }
     }
