+++ app/src/main/java/com/mimo/android/presentation/map/MapClusterBottomSheetDialogFragment.kt
@@ -9,11 +9,10 @@ import com.mimo.android.databinding.FragmentMapClusterBottomSheetDialogBinding
 import com.mimo.android.domain.model.MarkerData
 import com.mimo.android.domain.model.findMarkerIndex
 import com.mimo.android.presentation.base.BaseBottomSheetDialogFragment
-import com.mimo.android.presentation.map.adapter.MapClusterAdapter
+import com.mimo.android.presentation.component.adapter.MapClusterAdapter
 import com.mimo.android.presentation.util.getSizeY
 import com.mimo.android.presentation.videodetail.VideoDetailActivity
 import dagger.hilt.android.AndroidEntryPoint
-import timber.log.Timber
 
 @AndroidEntryPoint
 class MapClusterBottomSheetDialogFragment :
